refactor: search components
